+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "/home/justin/dev/dlib/ovid-bot/.venv/bin/python3.8"
+}
