+++ Pipfile
@@ -1,0 +1,17 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+pylint = "*"
+pep8 = "*"
+autopep8 = "*"
+
+[packages]
+tweepy = "*"
+apscheduler = "*"
+cltk = "*"
+
+[requires]
+python_version = "3.8"
