+++ app/twee.py
@@ -1,0 +1,17 @@
+import tweepy
+from os import environ
+
+
+#import Twitter credentials
+# from credentials import *
+
+# get Twitter creds from Heroku dashboard
+consumer_key = environ['CONSUMER_KEY']
+consumer_secret = environ['CONSUMER_SECRET']
+access_token = environ['ACCESS_KEY']
+access_token_secret = environ['ACCESS_SECRET']
+
+#access and authorize Twitter creds
+auth = tweepy.OAuthHandler(consumer_key, consumer_secret)
+auth.set_access_token(access_token, access_token_secret)
+api = tweepy.API(auth)
