--- requirements.txt
@@ -1,38 +1,0 @@
-APScheduler==3.6.3
-beautifulsoup4==4.8.2
-bs4==0.0.1
-certifi==2018.4.16
-chardet==3.0.4
-click==7.1.1
-cltk==0.1.116
-colorama==0.4.1
-decorator==4.4.1
-Flask==1.1.1
-gitdb2==2.0.3
-GitPython==2.1.10
-idna==2.9
-itsdangerous==1.1.0
-Jinja2==2.11.1
-MarkupSafe==1.1.1
-nltk==3.4.5
-oauthlib==3.1.0
-psutil==5.6.6
-PySocks==1.7.1
-pyte==0.8.0
-python-crfsuite==0.9.7
-pytz==2019.3
-pyuca==1.2
-PyYAML==5.3.1
-regex==2020.2.20
-requests==2.23.0
-requests-oauthlib==1.3.0
-six==1.11.0
-smmap2==2.0.3
-soupsieve==2.0
-thefuck==3.29
-tweepy==3.8.0
-tzlocal==2.0.0
-urllib3==1.25.8
-wcwidth==0.1.7
-Werkzeug==1.0.0
-Whoosh==2.7.4
